renamed weapon manager
- renamed to WeaponManagerComponent
