+++ Components/Scenes/weapon_manager_component.tscn
@@ -1,0 +1,6 @@
+[gd_scene load_steps=2 format=3 uid="uid://xodslpw7qcmg"]
+
+[ext_resource type="Script" path="res://Components/Scripts/WeaponManagerComponent.cs" id="1_aq2tq"]
+
+[node name="WeaponManagerComponent" type="Node2D"]
+script = ExtResource("1_aq2tq")
